Restructuring of project
- Move definitions.py to /definitions folder
- Update path in all files using definitions.py
- Add anti-pattern markdown files
