+++ sqleyes/definitions/anti-patterns/ambiguous_groups.md
@@ -1,0 +1,1 @@
+Hello Ambiguous Groups
