+++ sqleyes/definitions/anti-patterns/implicit_colums.md
@@ -1,0 +1,1 @@
+Hello Implicit Columns
