+++ sqleyes/definitions/anti-patterns/poor_mans_search_engine.md
@@ -1,0 +1,1 @@
+Hello Poor Man's Search Engine
