+++ sqleyes/definitions/anti-patterns/random_selection.md
@@ -1,0 +1,1 @@
+Hello Random Selection
