+++ sqleyes/detector/antipatterns/ambiguous_groups.py
@@ -2,7 +2,7 @@
 import re
 
 from sqleyes.detector.antipatterns.abstract_base_class import AbstractDetector
-from sqleyes.detector.definitions import DEFINITIONS
+from sqleyes.definitions.definitions import DEFINITIONS
 from sqleyes.detector.detector_output import DetectorOutput
 from sqleyes.utils.query_functions import (check_single_value_rule,
                                            get_columns_from_group_by_statement,
