+++ sqleyes/detector/antipatterns/random_selection.py
@@ -1,7 +1,7 @@
 """Random Selection anti-pattern detector class"""
 import re
 from sqleyes.detector.antipatterns.abstract_base_class import AbstractDetector
-from sqleyes.detector.definitions import DEFINITIONS
+from sqleyes.definitions.definitions import DEFINITIONS
 from sqleyes.detector.detector_output import DetectorOutput
